Split legacy and C2 functionality in two libs and add verify method in both
