+++ src/utils.ts
@@ -1,3 +1,6 @@
+import { sha256 } from 'js-sha256';
+import fs from "fs";
+
 export class utils {
 
     /**
@@ -9,5 +12,17 @@ export class utils {
             return false;
         }
         return true;
+    }
+
+    /**
+     * 
+     * @param agnosticData Either directly a binary file or a filepath to the file that should be hashed
+     * @param isBinaryInput Flag to check  the type provided in agnosticData-parm
+     * @returns 
+     */
+    public static hash(agnosticData: any, isBinaryInput: boolean): string {
+        const buffer: Buffer = !isBinaryInput ? fs.readFileSync(agnosticData) : agnosticData
+        const hash: string = sha256(buffer)
+        return hash
     }
 }
